production/Leonid_Smirnov_todoapp/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.7.2
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.1
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
Faker==8.12.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.1.2
openapi-codec==1.3.2
packaging==21.0
platformdirs==2.2.0
promise==2.3
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
